+++ mycodo/mycodo_flask/utils/utils_pid.py
@@ -6,6 +6,7 @@ from flask import flash
 from flask_babel import gettext
 
 from mycodo.config_translations import TRANSLATIONS
+from mycodo.databases.models import CustomController
 from mycodo.databases.models import DeviceMeasurements
 from mycodo.databases.models import Input
 from mycodo.databases.models import Output
@@ -334,19 +335,18 @@ def has_required_pid_values(pid_id, messages):
         PID.unique_id == pid_id).first()
 
     if not pid.measurement:
-        messages["error"].append(gettext(
-            "A valid Measurement is required"))
+        messages["error"].append("A valid Measurement is required")
     else:
         device_unique_id = pid.measurement.split(',')[0]
         input_dev = Input.query.filter(
             Input.unique_id == device_unique_id).first()
-        if not input_dev:
-            messages["error"].append(gettext(
-                "A valid Measurement is required"))
+        function = CustomController.query.filter(
+            CustomController.unique_id == device_unique_id).first()
+        if not input_dev and not function:
+            messages["error"].append("A valid controller is required")
 
     if not pid.raise_output_id and not pid.lower_output_id:
-        messages["error"].append(gettext(
-            "A Raise Output and/or a Lower Output is ""required"))
+        messages["error"].append("A Raise Output and/or a Lower Output is required")
 
     return messages